BLAST setup for populator and web server
Added a download step for NCBI BLAST in both Dockerfiles.
When the populator runs, if genes.fasta is present, it will copy the fasta file to a 'blast_db' folder in the content directory and make a BLAST database.
The function in the Flask app works, although potentially permissions will have to be changed to allow it to be run by a user.