Fixed plot bug when BLAST search is changed
The origin of the bug was due to a replot being triggered when on the BLAST tab. However, because the Search tab wasn't active, the size of the graph div element was 'auto', causing everything to fail.

This was fixed within the plot function, which will now halt execution if SVG width comes back as 'auto'.

When the Search tab is selected, a replot is now triggered, to allow for deselection due to changes in the BLAST tab to take effect.

In solving this problem, I also solved a bug that caused an error if the postcheckboxchange function was given a blank 'names' field.