Added checks for the 'BLAST Hits' reserved name
BLAST Hits is used by the system to send matching BLAST hits from the backend to the frontend. Checks have now been added to check for the presence of the name in the records and groups and fail if it is present.

Although there is probably a better way of sending the information without having to restrict the use of this name, using the name simplifies the frontend code, as 'BLAST Hits' can be treated like any other group.